Module tool::sequence

A bunch of traits and functions for working with sequences.

Modules

prelude
Useful functions exported by tool::sequence.

Traits

Empty
A sequence with no items.

First
A sequence with a first item.

Pair
A sequence with two items.

Second
A sequence with a second item.

Singleton
A sequence with one item.

Third
A sequence with a third item.

Triple
A sequence with three items.

Functions

first
Get the first element of a sequence with at least one element.

flip
Flip a length two tuple.

second
Get the second element of a sequence with at least to elements

third
Get the third element of a sequence with at least three elements.